refactor: Move user state out from auth
