+++ frontend/middleware/check-auth.js
@@ -1,5 +1,6 @@
 export default async function ({ store }) {
   if (!store.getters['auth/isAuthenticated'] || !store.getters['auth/getUsername']) {
     await store.dispatch('auth/initAuth')
+    store.dispatch('userState/init', store.getters['auth/getUserId'])
   }
 }
