+++ frontend/store/auth.js
@@ -1,12 +1,8 @@
-import moment from 'moment'
-
 export const state = () => ({
   username: null,
   id: null,
   isAuthenticated: false,
-  isStaff: false,
-  restingUserId: null,
-  restingEndTime: null
+  isStaff: false
 })
 
 export const mutations = {
@@ -24,14 +20,6 @@ export const mutations = {
   },
   setIsStaff(state, isStaff) {
     state.isStaff = isStaff
-  },
-  setRestingPeriod(state, endTime) {
-    state.restingUserId = state.id
-    state.restingEndTime = endTime
-  },
-  clearRestingPeriod(state) {
-    state.restingUserId = null
-    state.restingEndTime = null
   }
 }
 
@@ -47,37 +35,10 @@ export const getters = {
   },
   isStaff(state) {
     return state.isStaff
-  },
-  getRestingUserId(state) {
-    return state.restingUserId
-  },
-  getRestingEndTime(state) {
-    return state.restingEndTime
   }
 }
 
 export const actions = {
-  setRestingPeriod({ commit }) {
-    const startTime = new Date()
-    const endTime = moment(startTime).add(5, 'm').format('ddd, DD-MM-YYYY HH:mm:ss')
-    commit('setRestingPeriod', endTime)
-  },
-  getRestingPeriod({ commit, getters }) {
-    const currentTime = new Date()
-    const restingEndTime = getters.getRestingEndTime
-
-    if (currentTime >= restingEndTime) {
-      commit('clearRestingPeriod')
-    }
-
-    const currentUserId = getters.getUserId
-    const restingUserId = getters.getRestingUserId
-
-    if (restingUserId !== null && currentUserId === restingUserId) {
-      return moment(restingEndTime, 'ddd, DD-MM-YYYY HH:mm:ss').toDate()
-    }
-    return null
-  },
   async authenticateUser({ commit }, authData) {
     try {
       await this.$services.auth.login(authData.username, authData.password)
